major update 1
New and rewritten scripts for most of the systems.
Completely new systems (operations.js and notify.js/notify.css).
Several new and improved features for old systems.
A demonstation story file has been added.
Old scripts and scripts that haven't been updates yet can still be found
in the 'old' directory.
Documentation is still in the works.

This is the first of three updates as I refactor and improve everything
here.