the new simple inventory is here
it works a little different and isn't backward compatible with old code.
but it's so, so much better, i promise.

cleaned up some minor errors in a few other files, mostly style issues.

updated the demo and souce code files to include a section on simple
inventory.

docs are coming soon